+++ src/main/java/janggi/JanggiApplication.java
@@ -4,18 +4,24 @@ import janggi.config.AppConfig;
 import janggi.controller.JanggiController;
 import janggi.infra.ConnectionProvider;
 import janggi.infra.DbProperties;
+import janggi.infra.PersistenceConfig;
 
 public class JanggiApplication {
 
     public static void main(String[] args) {
         DbProperties dbProperties = new DbProperties();
-        AppConfig appConfig = new AppConfig(dbProperties);
+        PersistenceConfig persistenceConfig = new PersistenceConfig(dbProperties);
+        AppConfig appConfig = new AppConfig(persistenceConfig);
 
-        ConnectionProvider provider = appConfig.connectionProvider();
-        Runtime.getRuntime()
-                .addShutdownHook(new Thread(provider::dispose));
+        ConnectionProvider provider = persistenceConfig.connectionProvider();
+        registerShutdownHook(provider);
 
         JanggiController controller = appConfig.janggiController();
         controller.run();
+    }
+
+    private static void registerShutdownHook(ConnectionProvider connectionProvider) {
+        Runtime.getRuntime()
+                .addShutdownHook(new Thread(connectionProvider::dispose));
     }
 }
