+++ src/test/java/janggi/model/board/PlayingBoardTest.java
