+++ src/test/java/janggi/model/movement/StraightMovementTest.java
