+++ src/test/java/janggi/model/movement/palace/PalaceAdjacentMovementTest.java
