+++ src/test/java/janggi/model/movement/pattern/MovementPatternTest.java
