+++ src/test/java/janggi/model/movement/pattern/MovementPatternsTest.java
