+++ src/test/java/janggi/model/palace/PalaceTest.java
