+++ src/test/java/janggi/model/palace/PalacesTest.java
