+++ src/test/java/janggi/model/piece/ByeongTest.java
