+++ src/test/java/janggi/model/position/absolute/ColumnTest.java
