+++ src/test/java/janggi/model/position/absolute/PositionPathTest.java
