+++ src/test/java/janggi/model/position/absolute/PositionTest.java
