+++ src/test/java/janggi/model/position/absolute/RowTest.java
