+++ src/test/java/janggi/model/position/absolute/UndirectedLineTest.java
