+++ src/test/java/janggi/service/TestGameDao.java
@@ -68,3 +68,4 @@ public class TestGameDao implements GameDao {
         turnByGameId.put(gameId, nextTurn);
     }
 }
+
